feat : code improvement
