+++ dbms/memtablemanagement/memtablemanagement.go
@@ -10,3 +10,13 @@ type DatabaseMemtableManagement struct {
 	stateManagement   *statemanagement.DatabaseManagementStateManagement
 	stateGetOperation statemanagement.DatabaseManagementStateGetOperation
 }
+
+func NewMemtableManagement(
+	storage storage.StorageDirectories,
+	stateManagement *statemanagement.DatabaseManagementStateManagement) *DatabaseMemtableManagement {
+	return &DatabaseMemtableManagement{
+		storageDir:        storage,
+		stateManagement:   stateManagement,
+		stateGetOperation: stateManagement,
+	}
+}
